pyproject.toml
[project]
name = "openvalley-api"
version = "0.1.0"
description = "Warren Community Intelligence - AI agent for exploring community data"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ag-ui-protocol>=0.1.10",
    "beautifulsoup4>=4.12.0",
    "fastapi>=0.128.0",
    "geoalchemy2>=0.18.1",
    "google-api-python-client>=2.155.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "httpx>=0.28.1",
    "logfire[fastapi]>=4.16.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-ai>=1.39.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
    "pgvector>=0.3.0",
    "websockets<15.0,>=13.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
